requirements.txt
click==6.7
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.3
Flask==0.12
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
python-dateutil==2.6.1
six==1.10.0
SQLAlchemy==1.1.6
Werkzeug==0.12.2
